refactor: use Didot to abstract template engine implementation
