+++ DubUrl.Schema/DubUrl.Schema.csproj
@@ -10,7 +10,7 @@
 	</ItemGroup>
 
 	<ItemGroup>
-	  <PackageReference Include="Didot.Core" Version="0.27.10" />
+	  <PackageReference Include="Didot.Core" Version="0.31.0" />
 	</ItemGroup>
 
 	<PropertyGroup>
