+++ DubUrl.Testing/DubUrl.Testing.csproj
@@ -1,6 +1,5 @@
 <Project Sdk="Microsoft.NET.Sdk">
   <ItemGroup>
-    <PackageReference Include="Domemtech.StringTemplate4" Version="4.3.0" />
     <PackageReference Include="DuckDB.NET.Bindings.Full" Version="1.2.1" />
     <PackageReference Include="DuckDB.NET.Data" Version="1.2.1" />
     <PackageReference Include="Microsoft.Data.SqlClient" Version="6.0.2" />
