--- DubUrl.Testing/Querying/Templating/SqlRendererTest.cs
@@ -1,15 +1,0 @@
-﻿using DubUrl.Querying.Templating;
-using NUnit.Framework;
-using System;
-using System.Collections.Generic;
-using System.Globalization;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace DubUrl.Testing.Querying.Templating;
-
-public class SqlRendererTest
-{
-    
-}
